Continuous Graph Update:
Implemented continuous updating of the graph functionality. Users no longer need to manually click multiple times to update the graph. Now, the graph updates seamlessly, providing real-time visualization of stock data.
The graph now displays a line chart with the y-axis representing the stock's top_ask_price and the x-axis representing the stock's timestamp. This enhancement offers users a more intuitive and dynamic experience while analyzing stock data.
Duplicate Data Removal:

Addressed the issue of duplicate data that was previously observed. Duplicate data is now automatically removed or disregarded, ensuring data accuracy and integrity in the system.
Users will no longer encounter redundancy in the displayed data, leading to a cleaner and more reliable analysis of stock information.
These updates aim to enhance the usability and reliability of the system, providing users with a smoother experience while utilizing the graph functionality and ensuring the accuracy of the displayed data.